refactor: rename Number -> Numbers, NumbersValidator에서 static 메소드 제거
