+++ src/main/java/baseball/controller/PlayerController.java
@@ -1,6 +1,6 @@
 package baseball.controller;
 
-import baseball.domain.number.Numbers;
+import baseball.domain.numbers.Numbers;
 import baseball.domain.player.Player;
 import baseball.domain.result.Result;
 import baseball.view.BaseBallView;
