+++ src/main/java/baseball/domain/numbers/NumbersConstant.java
@@ -1,12 +1,12 @@
-package baseball.domain.number;
+package baseball.domain.numbers;
 
-public class NumberConstant {
+public final class NumbersConstant {
     public static final Integer MAX_NUMBER = 9;
     public static final Integer MIN_NUMBER = 1;
     public static final Integer NUMBER_COUNT = 3;
     public static final Integer MIN_INDEX = 0;
     public static final Integer MAX_INDEX = NUMBER_COUNT - 1;
 
-    private NumberConstant() {
+    private NumbersConstant() {
     }
 }
